+++ api/v2/types_utils.go
@@ -1,30 +1,12 @@
 package v2
 
 import (
-	"context"
-	"strconv"
-
-	"github.com/google/go-cmp/cmp"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/event"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
 )
 
 const (
 	// FinalizerName is the finalizer name used by this controller.
 	FinalizerName = "firewall.metal-stack.io/firewall-controller-manager"
-	// ReconcileAnnotation can be used to trigger a reconciliation of a resource managed by a controller.
-	// The value of the annotation does not matter, the controller will cleanup the annotation automatically and trigger a reconciliation of the resource.
-	ReconcileAnnotation = "firewall.metal-stack.io/reconcile"
-	// MaintenanceAnnotation can be used to trigger a maintenance reconciliation for which a controller might have special behavior.
-	// The value of the annotation does not matter, the controller will cleanup the annotation automatically.
-	MaintenanceAnnotation = "firewall.metal-stack.io/maintain"
-	// RollSetAnnotation can be used to trigger a rolling update of a firewall deployment.
-	// The value of the annotation needs to be true otherwise the controller will ignore it.
-	RollSetAnnotation = "firewall.metal-stack.io/roll-set"
-	// RevisionAnnotation stores the revision number of a resource.
-	RevisionAnnotation = "firewall.metal-stack.io/revision"
 )
 
 // ConditionStatus is the status of a condition.
@@ -114,122 +96,4 @@ func (cs Conditions) filterOutCondition(t ConditionType) Conditions {
 		newConditions = append(newConditions, c)
 	}
 	return newConditions
-}
-
-// IsAnnotationPresent returns true if the given object has an annotation with a given
-// key, the value of this annotation does not matter.
-func IsAnnotationPresent(o client.Object, key string) bool {
-	_, ok := o.GetAnnotations()[key]
-	return ok
-}
-
-// IsAnnotationTrue returns true if the given object has an annotation with a given
-// key and the value of this annotation is a true boolean.
-func IsAnnotationTrue(o client.Object, key string) bool {
-	enabled, err := strconv.ParseBool(o.GetAnnotations()[key])
-	return err == nil && enabled
-}
-
-// IsAnnotationFalse returns true if the given object has an annotation with a given
-// key and the value of this annotation is a false boolean.
-func IsAnnotationFalse(o client.Object, key string) bool {
-	enabled, err := strconv.ParseBool(o.GetAnnotations()[key])
-	return err == nil && !enabled
-}
-
-// RemoveAnnotation removes an annotation by a given key from an object if present by updating it with the given client.
-func RemoveAnnotation(ctx context.Context, c client.Client, o client.Object, key string) error {
-	annotations := o.GetAnnotations()
-
-	if annotations == nil {
-		return nil
-	}
-
-	_, ok := annotations[key]
-	if !ok {
-		return nil
-	}
-
-	delete(annotations, key)
-
-	o.SetAnnotations(annotations)
-
-	err := c.Update(ctx, o)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func annotationWasRemoved(update event.UpdateEvent, annotation string) bool {
-	if cmp.Equal(update.ObjectOld.GetAnnotations(), update.ObjectNew.GetAnnotations()) {
-		return false
-	}
-
-	var (
-		_, o = update.ObjectOld.GetAnnotations()[annotation]
-		_, n = update.ObjectNew.GetAnnotations()[annotation]
-	)
-
-	if n {
-		return false
-	}
-
-	if !o {
-		return false
-	}
-
-	return o && !n
-}
-
-func annotationWasAdded(update event.UpdateEvent, annotation string) bool {
-	if cmp.Equal(update.ObjectOld.GetAnnotations(), update.ObjectNew.GetAnnotations()) {
-		return false
-	}
-
-	var (
-		_, o = update.ObjectOld.GetAnnotations()[annotation]
-		_, n = update.ObjectNew.GetAnnotations()[annotation]
-	)
-
-	if o {
-		return false
-	}
-
-	if !n {
-		return false
-	}
-
-	return !o && n
-}
-
-// AnnotationAddedPredicate returns a predicate when the given annotation key was added.
-func AnnotationAddedPredicate(annotation string) predicate.Funcs {
-	return predicate.Funcs{
-		CreateFunc: func(ce event.CreateEvent) bool {
-			return false
-		},
-		UpdateFunc: func(update event.UpdateEvent) bool {
-			return annotationWasAdded(update, annotation)
-		},
-		DeleteFunc: func(de event.DeleteEvent) bool {
-			return false
-		},
-	}
-}
-
-// AnnotationRemovedPredicate returns a predicate when the given annotation key was removed.
-func AnnotationRemovedPredicate(annotation string) predicate.Funcs {
-	return predicate.Funcs{
-		CreateFunc: func(ce event.CreateEvent) bool {
-			return false
-		},
-		UpdateFunc: func(update event.UpdateEvent) bool {
-			return annotationWasRemoved(update, annotation)
-		},
-		DeleteFunc: func(de event.DeleteEvent) bool {
-			return false
-		},
-	}
 }
